move: ErrorResponse 이동 ( root/response -> root/exception )
